move isCollapsed
